requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
discord.py==2.5.2
frozenlist==1.5.0
idna==3.10
multidict==6.3.1
propcache==0.3.1
yarl==1.18.3
pymongo[srv]==3.12